Change usages of TypeInfo back to Type
Back when we were preparing the codebase for .NET Standard 1.0 (IIRC), we had to convert most instances of Type to instances of TypeInfo where Type did not support the required functionality (generally type metadata/reflection).

Now that we're firmly on .NET Standard 2.0 it's time to undo this change. Conversion between the two is not free and shows up during serialization benchmarks (around 2.7% inclusive time on serialization benchmarks)